+++ spring/src/test/java/com/influxdb/spring/influx/InfluxDB2AutoConfigurationReactiveTest.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.influxdb.spring.influx;
+
+import com.influxdb.client.InfluxDBClient;
+import com.influxdb.client.reactive.InfluxDBClientReactive;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.platform.runner.JUnitPlatform;
+import org.junit.runner.RunWith;
+import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.runner.ApplicationContextRunner;
+
+/**
+ * Tests for {@link InfluxDB2AutoConfiguration}.
+ *
+ * @author Jakub Bednar (bednar@github) (07/05/2019 12:59)
+ */
+@RunWith(JUnitPlatform.class)
+class InfluxDB2AutoConfigurationReactiveTest {
+
+    private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
+            .withConfiguration(AutoConfigurations.of(InfluxDB2AutoConfigurationReactive.class));
+
+
+    @Test
+    public void influxDBClientReactive() {
+        this.contextRunner.withPropertyValues("influx.url=http://localhost:8086/")
+                .run(((context) -> Assertions.assertThat(context.getBeansOfType(InfluxDBClientReactive.class))
+                        .hasSize(1)))
+                .run(((context) -> Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class))
+                        .hasSize(0)));
+    }
+}